zad1/requirements.txt
pyhamcrest
parameterized
coverage